move searchbox component out of search box folder because it's no longer needed
